% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_trunc_ndist}
\alias{rcpp_trunc_ndist}
\title{rcpp_trunc_ndist}
\usage{
rcpp_trunc_ndist(len, sd)
}
\arguments{
\item{len}{Number of elements to be simulated}

\item{sd}{Standard deviation}
}
\value{
A vector of truncated normally distributed values
}
\description{
Truncated normal distribution (mean 1, respective upper and lower limits of
0 and 2).
}

