% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\docType{package}
\name{tnorm-package}
\alias{tnorm}
\alias{tnorm-package}
\title{Fast Truncated Normal Distribution}
\description{
\pkg{tnorm} package implements dtnorm and rtnorm functions, using an
algorithm described in Robert (1995). The source codes are based on
Jonathan Olmsted's RcppTN 0.1-8 and Christopher Jackson's msm 1.5.
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au>
}
\references{
Robert, C. (1995). \emph{Simulation of truncated normal variables},
\emph{Statistics and Computing} \bold{5}, 121--125. \cr
Olmsted, J. \url{https://github.com/olmjo/RcppTN}.
}
\keyword{package}

