% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnormn.R
\name{tnormn}
\alias{tnormn}
\title{tnormn}
\usage{
tnormn(n, sd, seed)
}
\arguments{
\item{n}{Number of truncated normally distributed observations to be
simulated}

\item{sd}{Standard deviation of truncated normal distribution}

\item{seed}{Random seed}
}
\value{
A vector of n truncated normally distributed values
}
\description{
Generates a truncated normal distribution
}
\examples{
x <- tnormn (n=100, sd=0.1)

}

