% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncnorm.R
\docType{package}
\name{truncnorm}
\alias{truncnorm}
\alias{truncnorm-package}
\title{truncnorm}
\description{
Generates truncated normal distributions. Current best R method is
\code{msm::rtnorm}, but this is entirely R-based and therefore s l o w.
\code{trunc-norm} is \code{Rcpp}-based and much faster.
}
\section{Functions}{

\tabular{ll}{
\code{junk}\tab description
}
}

