% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvtnorm.R
\name{rmvtnorm}
\alias{rmvtnorm}
\title{Generate truncated normal samples.}
\usage{
rmvtnorm(
  n,
  mean = c(0, 0),
  cov = diag(length(mean)),
  initial = rep(0, length(mean)),
  F = NULL,
  g = NULL,
  burn = 10L
)
}
\arguments{
\item{n}{number of samples.}

\item{mean}{mean vector.}

\item{cov}{variance covariance matrix.}

\item{initial}{an initial value for the Markov chain.}

\item{F}{constraint matrix defining linear inequalities.}

\item{g}{constraint vector defining linear inequalities.}

\item{burn}{number of burn-in iterations before collecting samples.}
}
\value{
A matrix with each row corresponding to a sample.
}
\description{
Generate truncated normal samples using Hamiltonian Monte Carlo technique.
}
\details{
This is a modified version of \code{tmg} package in CRAN. The samples will satisfy
the following inequality \deqn{F \times x + g >= 0} where \eqn{x} is a column vector
of the generated sample.

No check for symmetry is performed on covariance. The initial value must
strictly satisfy the inequality constraints, although the generated samples
only satisfy them weakly.
}
\examples{
plot(rmvtnorm(100))

plot(rmvtnorm(100, F = diag(2), g = c(1, -1), initial = c(0, 2)))

plot(rmvtnorm(100, mean = c(1, 1), cov = matrix(c(1, 0.5, 0.5, 1), 2), F =
diag(2), g = c(1, -1), initial = c(0, 2)))

}
