% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\docType{package}
\name{tnorm}
\alias{tnorm}
\alias{tnorm-package}
\title{tnorm}
\description{
Generates truncated normal distributions. Current best R method is
\code{msm::rtnorm}, but this is entirely R-based and therefore s l o w.
\code{trunc-norm} is \code{Rcpp}-based and much faster.
}
\section{Functions}{

\tabular{ll}{
\code{tnormn}\tab The sole function; used to generate truncated normal
distributions
}
}

