% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\docType{package}
\name{tnorm}
\alias{tnorm}
\alias{tnorm-package}
\title{Truncated Normal Distribution}
\description{
Truncation normal distribution for density function, random number generator,
and cumulative distribution function in C++.
}
\references{
Robert, C. (1995). \emph{Simulation of truncated normal variables},
\emph{Statistics and Computing} \bold{5}, 121--125.
}
\author{
Yi-Shin Lin <yishin.lin@utas.edu.au>
}
\keyword{package}
